Use double instead of float for primitive

+++ source/primitive.cpp
@@ -12,23 +12,23 @@ Primitive::Primitive(std::shared_ptr<Geometry> geometry, std::shared_ptr<Materia
     assert(m_material != nullptr);
 }
 
-std::optional<GeometryHit> Primitive::geometry_raycast(const float3& origin, const float3& direction, float length) const {
+std::optional<GeometryHit> Primitive::geometry_raycast(const float3& origin, const float3& direction, double length) const {
     assert(isfinite(origin));
-    assert(equal(::length(direction), 1.f));
-    assert(length >= 0.f);
+    assert(equal(::length(direction), 1.0));
+    assert(length >= 0.0);
 
     float3 object_origin(point_transform(origin, m_inv_transform));
     float3 object_direction(normalize(direction * m_inv_transform));
 
     assert(isfinite(object_origin));
-    assert(equal(::length(object_direction), 1.f));
+    assert(equal(::length(object_direction), 1.0));
 
     std::optional<GeometryHit> object_result = m_geometry->raycast(object_origin, object_direction, length);
     if (object_result) {
         assert(isfinite(object_result->position));
-        assert(equal(::length(object_result->tangent), 1.f));
-        assert(equal(::length(object_result->bitangent), 1.f));
-        assert(equal(::length(object_result->normal), 1.f));
+        assert(equal(::length(object_result->tangent), 1.0));
+        assert(equal(::length(object_result->bitangent), 1.0));
+        assert(equal(::length(object_result->normal), 1.0));
         assert(object_result->distance >= 0 && object_result->distance <= length);
 
         float3 world_position = point_transform(object_result->position, m_transform);
@@ -37,10 +37,10 @@ std::optional<GeometryHit> Primitive::geometry_raycast(const float3& origin, con
         float3 world_normal = normalize(normal_transform(object_result->normal, m_inv_transform));
 
         assert(isfinite(world_position));
-        assert(equal(::length(world_tangent), 1.f));
-        assert(equal(::length(world_bitangent), 1.f));
-        assert(equal(::length(world_normal), 1.f));
-        assert(object_result->distance > 0.f && object_result->distance <= length);
+        assert(equal(::length(world_tangent), 1.0));
+        assert(equal(::length(world_bitangent), 1.0));
+        assert(equal(::length(world_normal), 1.0));
+        assert(object_result->distance > 0.0 && object_result->distance <= length);
 
         GeometryHit world_result;
         world_result.position = world_position;
@@ -55,15 +55,15 @@ std::optional<GeometryHit> Primitive::geometry_raycast(const float3& origin, con
 }
 
 GeometrySample Primitive::geometry_sample(const float2& random) const {
-    assert(random[0] >= 0.f && random[0] < 1.f);
-    assert(random[1] >= 0.f && random[1] < 1.f);
+    assert(random[0] >= 0.0 && random[0] < 1.0);
+    assert(random[1] >= 0.0 && random[1] < 1.0);
 
     GeometrySample object_space_sample = m_geometry->sample(random);
 
     assert(isfinite(object_space_sample.position));
-    assert(equal(length(object_space_sample.tangent), 1.f));
-    assert(equal(length(object_space_sample.bitangent), 1.f));
-    assert(equal(length(object_space_sample.normal), 1.f));
+    assert(equal(length(object_space_sample.tangent), 1.0));
+    assert(equal(length(object_space_sample.bitangent), 1.0));
+    assert(equal(length(object_space_sample.normal), 1.0));
 
     GeometrySample world_space_sample;
     world_space_sample.position = point_transform(object_space_sample.position, m_transform);
@@ -72,43 +72,43 @@ GeometrySample Primitive::geometry_sample(const float2& random) const {
     world_space_sample.normal = normalize(normal_transform(object_space_sample.normal, m_inv_transform));
 
     assert(isfinite(world_space_sample.position));
-    assert(equal(length(world_space_sample.tangent), 1.f));
-    assert(equal(length(world_space_sample.bitangent), 1.f));
-    assert(equal(length(world_space_sample.normal), 1.f));
+    assert(equal(length(world_space_sample.tangent), 1.0));
+    assert(equal(length(world_space_sample.bitangent), 1.0));
+    assert(equal(length(world_space_sample.normal), 1.0));
 
     return world_space_sample;
 }
 
-float Primitive::geometry_area() const {
-    float result = m_geometry->area();
+double Primitive::geometry_area() const {
+    double result = m_geometry->area();
 
     assert(std::isfinite(result));
-    assert(result >= 0.f);
+    assert(result >= 0.0);
 
     return result;
 }
 
 float3 Primitive::material_bsdf(float3& ingoing, const float3& outgoing, const float2& random) const {
-    assert(equal(length(outgoing), 1.f));
-    assert(random[0] >= 0.f && random[0] < 1.f);
-    assert(random[1] >= 0.f && random[1] < 1.f);
+    assert(equal(length(outgoing), 1.0));
+    assert(random[0] >= 0.0 && random[0] < 1.0);
+    assert(random[1] >= 0.0 && random[1] < 1.0);
 
     float3 result = m_material->bsdf(ingoing, outgoing, random);
 
-    assert(result.x >= 0.f && result.y >= 0.f && result.z >= 0.f);
-    assert(equal(length(ingoing), 1.f));
+    assert(result.x >= 0.0 && result.y >= 0.0 && result.z >= 0.0);
+    assert(equal(length(ingoing), 1.0));
 
     return result;
 }
 
 float3 Primitive::material_bsdf(const float3& ingoing, const float3& outgoing) const {
-    assert(equal(length(ingoing), 1.f));
-    assert(equal(length(outgoing), 1.f));
+    assert(equal(length(ingoing), 1.0));
+    assert(equal(length(outgoing), 1.0));
 
     float3 result = m_material->bsdf(ingoing, outgoing);
 
     assert(isfinite(result));
-    assert(result.x >= 0.f && result.y >= 0.f && result.z >= 0.f);
+    assert(result.x >= 0.0 && result.y >= 0.0 && result.z >= 0.0);
 
     return result;
 }
@@ -117,7 +117,7 @@ float3 Primitive::material_emissive() const {
     float3 result = m_material->emissive();
 
     assert(isfinite(result));
-    assert(result.r >= 0.f && result.g >= 0.f && result.b >= 0.f);
+    assert(result.r >= 0.0 && result.g >= 0.0 && result.b >= 0.0);
 
     return result;
 }
